Move the methods to metrics.py and import them
