+++ src/pal_mjlab/tasks/velocity/kangaroo/env_cfgs.py
@@ -1,7 +1,5 @@
 """PAL Robotics KANGAROO velocity tracking environment configurations."""
 
-import torch
-
 from mjlab.envs import ManagerBasedRlEnvCfg
 from mjlab.envs.mdp.actions import JointPositionActionCfg
 from mjlab.managers.event_manager import EventTermCfg
@@ -221,38 +219,12 @@ def pal_kangaroo_rough_env_cfg(play: bool = False) -> ManagerBasedRlEnvCfg:
         params={"sensor_name": self_collision_cfg.name},
     )
 
-    ## Metrics Calculations
-    def joint_velocity_magnitude(env, asset_cfg):
-        """L1 norm of joint velocities."""
-        return env.scene[asset_cfg.name].data.joint_vel.abs().sum(dim=-1)  # (num_envs,)
-
-    def joint_accelerations_magnitude(env, asset_cfg):
-        """L1 norm of joint velocities."""
-        return env.scene[asset_cfg.name].data.joint_acc.abs().sum(dim=-1)  # (num_envs,)
-
-    def joint_torques_magnitude(env, asset_cfg):
-        """L1 norm of joint torques."""
-        return env.scene[asset_cfg.name].data.actuator_force.abs().sum(dim=-1)  # (num_envs,)
-
-    def action_rate_l2(env) -> torch.Tensor:
-        """Penalize the rate of change of the actions using L2 squared kernel."""
-        return torch.sum(
-            torch.square(env.action_manager.action - env.action_manager.prev_action), dim=1
-        )
-    def action_acc_l2(env) -> torch.Tensor:
-        """Penalize the acceleration of the actions using L2 squared kernel."""
-        action_acc = (
-            env.action_manager.action
-            - 2 * env.action_manager.prev_action
-            + env.action_manager.prev_prev_action
-        )
-        return torch.sum(torch.square(action_acc), dim=1)
-
-    cfg.metrics = {"joint_vel_mag": MetricsTermCfg(func=joint_velocity_magnitude, params={"asset_cfg": SceneEntityCfg("robot", joint_names=(".*",))}),
-                "joint_acc_mag": MetricsTermCfg(func=joint_accelerations_magnitude, params={"asset_cfg": SceneEntityCfg("robot", joint_names=(".*",))}),
-                "joint_torque_mag": MetricsTermCfg(func=joint_torques_magnitude, params={"asset_cfg": SceneEntityCfg("robot", joint_names=(".*",))}),
-                "action_rate_l2": MetricsTermCfg(func=action_rate_l2, params={}),
-                "action_acc_l2": MetricsTermCfg(func=action_acc_l2, params={})}
+    ## Metrics
+    cfg.metrics = {"joint_vel_mag": MetricsTermCfg(func=mdp.joint_velocity_magnitude, params={"asset_cfg": SceneEntityCfg("robot", joint_names=(".*",))}),
+                "joint_acc_mag": MetricsTermCfg(func=mdp.joint_accelerations_magnitude, params={"asset_cfg": SceneEntityCfg("robot", joint_names=(".*",))}),
+                "joint_torque_mag": MetricsTermCfg(func=mdp.joint_torques_magnitude, params={"asset_cfg": SceneEntityCfg("robot", joint_names=(".*",))}),
+                "action_rate_l2": MetricsTermCfg(func=mdp.action_rate_l2, params={}),
+                "action_acc_l2": MetricsTermCfg(func=mdp.action_acc_l2, params={})}
 
     # # All except leg length joints
     # cfg.rewards["joint_accel"] = RewardTermCfg(
