+++ src/pal_mjlab/tasks/velocity/kangaroo/mdp/__init__.py
@@ -3,3 +3,4 @@ from mjlab.tasks.velocity.mdp import *  # noqa: F401, F403
 
 from .rewards import *  # noqa: F401, F403
 from .observations import *  # noqa: F401, F403
+from .metrics import *  # noqa: F401, F403
